chore: enhance FIDO2 build script and domain quality
- Removed redundant `toInt()` calls in `PasskeyCommand.Register.hashCode()` for improved efficiency.
- Resolved IDE warnings and detekt issues in `:feature:fido2/build.gradle.kts`, including KDoc syntax errors and unused property delegates.
- Migrated from deprecated `compose.runtime` to a stable, version-catalog-managed `libs.compose.runtime` dependency.
- Updated CHANGELOG.md to document these changes and their impact on code quality and maintainability.

+++ core/domain/src/commonMain/kotlin/com/chimali/core/domain/eventsourcing/passkey/PasskeyCommand.kt
@@ -42,7 +42,7 @@ sealed interface PasskeyCommand {
             result = 31 * result + credentialId.contentHashCode()
             result = 31 * result + publicKey.hashCode()
             result = 31 * result + aaguid.hashCode()
-            result = 31 * result + signCount.hashCode().toInt()
+            result = 31 * result + signCount.hashCode()
             return result
         }
     }
